Moved tests into the more general class
