+++ cassandra/src/main/java/se/tre/freki/storage/cassandra/CassandraStore.java
@@ -37,7 +37,6 @@ import com.datastax.driver.core.Row;
 import com.datastax.driver.core.Session;
 import com.google.common.base.Function;
 import com.google.common.base.Optional;
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
@@ -413,10 +412,6 @@ public class CassandraStore extends Store {
         }
 
         final Row row = metaRow.get();
-
-        if (Strings.isNullOrEmpty(row.getString("description"))) {
-          return Optional.absent();
-        }
 
         return Optional.of(
             LabelMeta.create(CassandraLabelId.fromLong(row.getLong("label_id")),
